requirements.txt
# Core dependencies only (no AI models)
# ~200MB download vs ~5GB+ for full requirements

PySide6>=6.6.0
PyOpenGL>=3.1.0
Pillow>=10.0.0
numpy>=1.24.0
aiohttp>=3.9.0
msgpack>=1.0.0

# Local inference (stable-diffusion.cpp)
stable-diffusion-cpp-python>=0.4.0
huggingface-hub>=0.20.0
